requirements.txt
openai-clip==1.0.1
lpips==0.1.4
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.2.0
scikit_learn==1.0.2
scipy==1.7.3
six==1.16.0
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
